# openAPI-spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce Backend API (Generated from Code)",
    "version": "1.0.0",
    "description": "API specification for the Telegram-based e-commerce platform, generated directly from the implemented backend code."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Base URL for the API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "A human-readable error message."
          }
        },
        "required": ["message"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "ObjectId", "description": "Unique identifier for the user." },
          "telegramId": { "type": "integer", "description": "Telegram user ID." },
          "username": { "type": "string", "description": "User's chosen username." },
          "role": { "type": "string", "enum": ["customer", "vendor", "admin"], "description": "User's role." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "telegramId", "username", "role"]
      },
      "UserRegistration": {
        "type": "object",
        "properties": {
          "telegramId": { "type": "integer", "description": "Telegram user ID." },
          "username": { "type": "string", "description": "User's chosen username." },
          "role": { "type": "string", "enum": ["customer", "vendor", "admin"], "default": "customer", "description": "User's role (defaults to customer)." }
        },
        "required": ["telegramId", "username"]
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "telegramId": { "type": "integer", "description": "Telegram user ID." }
        },
        "required": ["telegramId"]
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "description": "New username for the user." }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "token": { "type": "string", "description": "JWT authentication token." }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserRegistration" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "409": {
            "description": "Username or Telegram ID already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login a user",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserLogin" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "summary": "Get current user profile",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update current user profile",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}